Move from_yaml methods out of metadata.py and into their own file.
